refactor v2 object serialization
