+++ core/v2/serialization_test.go
@@ -56,7 +56,7 @@ func TestBlobKeyFromHeader(t *testing.T) {
 	assert.Equal(t, "b19d368345990c79744fe571fe99f427f35787b9383c55089fb5bd6a5c171bbc", blobKey.Hex())
 }
 
-func TestBatchHeaderHAsh(t *testing.T) {
+func TestBatchHeaderHash(t *testing.T) {
 	batchRoot := [32]byte{}
 	copy(batchRoot[:], []byte("1"))
 	batchHeader := &v2.BatchHeader{
@@ -68,4 +68,32 @@ func TestBatchHeaderHAsh(t *testing.T) {
 	assert.NoError(t, err)
 	// 0x891d0936da4627f445ef193aad63afb173409af9e775e292e4e35aff790a45e2 verified in solidity
 	assert.Equal(t, "891d0936da4627f445ef193aad63afb173409af9e775e292e4e35aff790a45e2", hex.EncodeToString(hash[:]))
+}
+
+func TestBlobCertHash(t *testing.T) {
+	data := codec.ConvertByPaddingEmptyByte(GETTYSBURG_ADDRESS_BYTES)
+	commitments, err := p.GetCommitments(data)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	blobCert := &v2.BlobCertificate{
+		BlobHeader: &v2.BlobHeader{
+			BlobVersion:     0,
+			BlobCommitments: commitments,
+			QuorumNumbers:   []core.QuorumID{0, 1},
+			PaymentMetadata: core.PaymentMetadata{
+				AccountID:         "0x123",
+				BinIndex:          5,
+				CumulativePayment: big.NewInt(100),
+			},
+			Signature: []byte{1, 2, 3},
+		},
+		RelayKeys: []v2.RelayKey{4, 5, 6},
+	}
+
+	hash, err := blobCert.Hash()
+	assert.NoError(t, err)
+	// 0xc4512b8702f69cb837fff50a93d3d28aada535b1f151b64db45859c3f5bb096a verified in solidity
+	assert.Equal(t, "c4512b8702f69cb837fff50a93d3d28aada535b1f151b64db45859c3f5bb096a", hex.EncodeToString(hash[:]))
 }
